Abstract KPI drop handling into a hook
JIRA: RAIL-4170
